requirements.txt
telethon==1.36.0
pandas==2.2.2
numpy>=1.26.0
pyyaml==6.0.2
flake8==7.1.1
pytest==8.3.3
pytest-cov==5.0.0
papermill==2.6.0
jupyter==1.1.1
emoji>=2.0.0
spacy
transformers>=4.30.0
datasets>=2.14.0
seqeval>=1.2.2
accelerate>=0.20.0
torch>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
shap>=0.42.0
lime>=0.2.0.1
evaluate>=0.4.0
tqdm>=4.65.0